Remove Ubuntu 14.04's ffmpeg support (#564)
The ffmpeg suport in tensorflow-io used to cover Ubuntu 14.04, 16.04, 18.04:
- Ubuntu 14.04: through libav9.20
- Ubuntu 16.04: through ffmpeg 2.8
- Ubuntu 18.04: through ffmpeg 3.4

As Ubuntu 14.04 has been EOL for some time, and we don't
use ubuntu 14.04 to build wheel files any more, it is time
to remove the Ubuntu 14.04's ffmpeg support in tensorflow-io.
